requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
pypdf
plotly
matplotlibnetworkx
